requirements.txt
asttokens==3.0.0
autopep8==2.3.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
executing==2.1.0
Flask==3.1.0
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy==1.13.0
mypy-extensions==1.0.0
nbqa==1.9.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
stack-data==0.6.3
tokenize_rt==6.1.0
tomli==2.2.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
